pyproject.toml
[project]
name = "oop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic-ai>=1.39.0",
    "pydantic-settings>=2.12.0",
]
